+++ packages/claims-controller/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new public method, `fetchClaimsConfigurations` to fetch the claims configuration from the Claims backend. ([#7109](https://github.com/MetaMask/core/pull/7109))
+- Added new states fields, `validSubmissionWindowDays` and `supportedNetworks` to the controller state. ([#7109](https://github.com/MetaMask/core/pull/7109))
+  - `validSubmissionWindowDays` - number of days the claim is valid for submission.
+  - `supportedNetworks` - supported networks for the claim submission.
+
 ## [0.1.0]
 
 ### Added